went back and looked at size of classes
The base employee class had a size of 88, while owner had the same
amount. This makes sense as it is a child class and adds no more data
members to the class. Waiter added 8 more bytes with two ints being
added on top of parent class size, tips and experience. Chef was the
largest at 120 bytes, which made sense.