more built in vector functions
Added the clear function to the vector to show that it can be cleared. I
feel like this will be useful down the road when I need to use vectors
as a non-persisant buffer or something along those lines. I imagine
vectors would be really good for online shopping cart itemization and
item tracking.