Had fun using cmath
Created a table that was not only a tiems table, but also aquared the
resulting value before printing. At first I used int as a data type and
this caused a table with -278237843 value to appear. I realized from the
size of the number and from teh sign, that I had an overflow. So, to fix
this I used a 'double' data type as the last data type to store before
printing and this fixed the problem!