vector array with indexing
Need to learn more about stack vs heap when dealing with large quantities.
For example an array of ints will overflow if tried using 10^7 elements.
Therefore I want to learn how to have very large elements that use vectors because
I know vectors are dynamic and dynamically store on the heap which is much larger
than the stack.